feat: Removing all deprecated features ("moving initative on dying", "moving initiative manually", "purging effects each turn", "purging effects on time increase").
'E's not pinin'! 'E's passed on! This feature is no more! It has ceased to be! 'E's expired and gone to meet 'its maker! 'E's a stiff! Bereft of life, 'e rests in peace! If you hadn't nailed 'it to the code 'e'd be pushing up the daisies! 'Is processes are now 'istory! 'E's off the browser! 'E's kicked the bucket, 'e's shuffled off 'is mortal coil, run down the curtain and joined the bleedin' choir invisible!! THIS IS AN EX-FEATURE!!