feat: My attempt at ABP using Rule Elements (hosted on a 'Feat').
Not quite complete, but, useful enough. (My players decided against ABP so I won't use it in the near term, but figured it might be useful for someone else, so I'm putting it in the Workbench.)
To use it, *do not* turn on ABP in Foundry, instead put this 'feat' in the Campaign Feats section on all characters you want to use these rules.)
Only lightly tested, but you probably need to remove and readd this feat after you level up.
Adjust the economy as much as you feel is needed. You don't need to remove runes from weapons and armor if you don't want to, these REs are implemented as Item Bonuses, so whichever of the rune and the character has the highest bonus is applied, *except* for Devastating Attacks which *does* stack with Striking Runes. (I used the Aroden icon as, well, I just discussed the campaign economy in the previous sentence, so, it felt appropriate. :)