fix: Fix skill action slugs not being added along to rolls
For skill actions where there is no system equivalent, the Workbench generates a roll like them, filling it with related data. The system itself also follows a similar format (for example, Flexible Catfolk has a bonus to `action:escape`, but no such macro exists in the system). Previously, if you used the skill action button on the sheet to escape, said bonus would not get added -- this was because the code was trying to call `this.slug` rather than `this.data.slug`, the result always ending up as `action:undefined`, which (obviously) doesn't match the `action:esacpe` predicate. This fixes this issue.