Fix URL for background image
The new actions dialog is different, don't apply anymore.
Still applies to the other one though.

It's not right if the foundry base URL isn't at the server's root.  It's
not right ever, but it only breaks in the non-root install.

The base URL for the style url() call is the foundry server base.
Doesn't need to be a "../" in front.  I.e., the URL is:

http://server.com/../systems/pf2e/assets/sheet/background.webp

The invalid "../" is removed by the css url function so it works.  But
for a non-root install, one might have:

http://server.com/fvtt/../systems/pf2e/assets/sheet/background.webp

And now the "../" isn't invalid, so it stays, and the URL is wrong.