Sort Basic Action Macro dialog by columns differently
This is much simpler than that existing code and doesn't result in huge css
file with thousands of classes.

Currently the css file is 7675 lines and 250 kB, almost as large as the
total mjs code.  After this change css files is 425 lines and only 8 kB.

The dialog template injects the number of BAM items as a CSS variable,
with style="...", in the div that contains the BAM action list.

Then the limited math that can be done in CSS itself is enough to figure
out how many rows and columns should be used for the grid.  This can be
done in just a single class, using CSS variables for the number of items
(from above) and the number of columns for the window width, which the
existing per-window width classes specify.

I probably wouldn't go up so such a huge number of columns.  Does anyone
actually want a window with 40 columns that stretches across two
monitors and then some?  Seems like 10 columns or less should look fine.