Filter the FollowTheExpert choices automatically from the target
If a target is selected when using FTE, filter the allowed skill choices
for the ChoiceSet to just those the target (i.e., the titular "expert")
has at expert or better proficiency.

Then add predicates to all the proficiency rank choices that check if
the selected skill is one of the skils the target had at that
proficiency rank.  This means only one of the choices should have a
predicate that passes, and with just one choice the dialog will be
skipped and the rank is effectively selected automatically.

Example choice would look like:
{
  "label": "PF2E.ProficiencyLevel2",
  "sort": 1,
  "value": 2,
  "predicate": {
    "or": [
      {
	"eq": [
	  "{item|flags.pf2e.rulesSelections.followTheExpertSkill}",
	  "intimidation"
	]
      }
    ]
  }
}

In this case, the target has just one still at Expert level,
intimidation.  If that was chosen as the skill, then the predicate
passes and the choice is allowed.  If another skill was chosen, then
expert is not allowed.

If there is no target, then it just adds the effect the same as it did
before.