Allow FollowTheExpert at a lower proficiency level via actor flag
Instead of always using expert as the minimum proficiency level needed
to follow, look at the actor flag "pf2e.followTheExpert.minimum" for the
value.  The default is 2 if this flag isn't defined.  This flag is on
the target, i.e. the "expert", not the follower.

Thaumaturge Regalia initiate benefit sets this to 1, to allow using FTE
when trained.  Maybe there are or will be other things that do this too.