Switch FollowTheExpert over to a typescript Action
This implementation should do the exact same thing as what the macro
did.

An improvement is that it's now correctly listed on the Exploration tab
since Actions can provide the traits used for sorting while macros can
not.

Delete the old macro.

There are not yet any examples of typescript actions other than those
already created by the pf2e system.  What I've done is create a new
class that implements the Action interface that the existing pf2e
actions use.

I'd have liked to inherit from SimpleAction, but it's not clear how to
do this since the full source of pf2e isn't available (the declarations
files are not sufficient) and the system doesn't provide the
SimpleAction and SimpleActionVariant classes anywhere global.

It wouldn't save any code, since most of what SimpleAction does isn't
needed here and there is a bunch of boilerplate to write for the
non-optional variant support that isn't used.