feat: BAM: Get rid of eval for Actions with a statistic
Add a new `options` field to the MacroAction object, with options for an
Action `use()` call.  Most of eval'ed string actions left are `.use()`
calls with a `statistic` option supplied.  This can now be done with action
as the Action object and an options field with the statistic.

The special processing of strings with ".use(" can go away now, as there
are none left.