feat: BAM: Remove last eval'ed string
The aid macro from pf2e-macro's module is the only one left.  Replace it
with an arrow function.  I can't test this as that module is no longer
installable.  But I think this will work.

The code and option to have the action be just a string is now gone.