feat: BAM: Wrap macros in a class.
This gets rid of all the BAM actions that were specified as an array of
strings.  A new class that implements ActionVariant is created, it works by
executing a macro for the use() method, with the macro name supplied during
construction.

Typescript's type checking then caught that "event" is not an argument to
Macro.execute().  The event needs to be in an object along with anything else
to be placed into the scope of the macro.  Which is now done.  The allows
the shift click on the BAM button to control the slow roll dialog.

The existing code would download _all_ documents in the compendium the
macro was in.  It will now specify a name in the query and foundry will only
download the macro to be executed.

It would also create a new macro from the existing one, which appears to be
unnecessary.