fix: BAM: Use lowercase skill names
Every use of the skill name, except one, turned it into a lowercase
skill slug.  And the one that didn't was a bug!

Just make it lowercase in the action data to start with.

The bug was in the getSkills() call, but it didn't matter since none of the
actions that use it have MAP.  I.e., it would try to look up
`actor.skills.Crafting` and fail for the repair action, but the value would
never be used since only the MAP variant code used it and repair has no MAP.
The same for Aid and Raise a Shield, the only other call sites.