Fix the macro wrapper for V12
This also makes it more efficient.  It was downloading every single
macro in the compendium, then finding the selected one to copy and
execute.

getDocuments can search by name, it's an indexed field, so use that to
only download the selected macro.  And only make a copy of the macro if
it can't be executed.  The default permissions will allow it to be
executed without a copy.