chore: Delete BAM code for variant non-object actions
Almost every action is now an object implementing the Action interface from
the pf2e system.  Only four are not:  the Aid macro from the activemacros
module, Craft, Raise a Shield, and Repair.

These don't use MAP, so none of the code relating to creating a variant
with a MAP modifier added is needed.

Get rid of <Function> cast too.  It's possible to avoid it by checking for the
remaining possible type, undefined, before calling the property.  Typescript
will deduce it must be a function.