fix: Unify apply persistent damage/healing, add roll options
Now that there are REs that can trigger on damage-received and
healing-received, having the right roll options set when taking the
damage/healing matters.  The auto apply feature wasn't getting any of
them.  Fix that.

Combine the persistent healing and damage functions, since a lot of the
code is shared.

Use the heal flavor i18n identifier from PF2e system, so that the translations
don't need to also be added to the module and kept up to date with every
language in the pf2e system.

Move the hook to preCreateChatMessage.  This way it only runs on the
single client of the user who is sending the notification, which is
typically the GM.  The system already figures out a user to send the
damage roll, so we can just have the same user apply the damage too.