fix: Make BAM work with Dorako UI
The way it did the background color and the choice of selectors made it
problematic to theme BAM.

Instead of adding a style background directly to the buttons, give them
data properies with the skill rank.  Then use that in the css to select
colors.  Which now use var names from the pf2e css that match the abilities
tab on the char sheet, instead of copying the color hex codes.

With the foreground and background color in the same place, it's possible
to override neither or both.  But overriding one, so that a light theme
foreground is combined with a dark theme background, shouldn't happen.