feat: Add option for Recall Knowledge breakdown style
This controls the style of the roll breakdown shown by the Recall Knowledge
macro.  It can show no modifiers, all modifiers, or only the extra modifiers
beyond the standard skill and proficiency modifiers that are always present.

It's added to the settings of the BAM macro settings.  Maybe both RK and BAM
settings should go on a new sub-menu?