fix: Don't hide crafting item information
The craft *action* chat card has some information on the item being crafted that
re-uses the html from an *item* chat card.  The collapse item chat card feature
was hiding this bit inside the action card.  But because it wasn't an item chat
card there was no way to unhide it.

Changing the selector to not hide item content inside flavor text will avoid
this.