fix: Fix logic for alt skills with feat in BAM
It was testing if the actor had any alt skill and then if the actor had any alt
feat.  So being trained in Nature and having the Chirurgeon feat would count for
using Treat Wounds.

Change this so after checking for the alt skill it then checks if the
corresponding alt feat is present.  So being trained in Nature requires Natural
Medicine to be present.

And use some() instead of find().