feature: Put item collapse trigger on entire header row
The collapse click trigger for item chat cards was only the item name.  The icon
to left of the name didn't trigger, nor did the eye icon on the right.  The eye
was especially glaring as only the eye icon is the trigger for other types of
collapsed chat cards.

Also improve the efficiency of the handler.  It was installing a click handler
for any click anywhere in any chat card.  The handler will query the event and
card and figure out if it was a click in the title of an item card.

Since we already figured out of the card was an item card, we can add a handler
for only that type of card.  And it can be added to the header row only and not
the entire card.