fix: Dirge of Doom effect didn't apply Frightened
Needed inMemoryOnly.  Also restict frightened from being removed when in the
aura.  Allow duplicates so an already frightened creature gaining the effect
will gain the "frightened that doesn't go away" from the effect in addition the
the frightened that does go away that is already has.

And clean up some old fields.