fix: Make BAM work on v12/13 and fix non-tabview section headers
Fixes the issues with the code on v12 without duplicating the entire
dialog.

Remove the ts-expect-error, which was hiding the problem that broke v12.

Then fix all the other typescript errors.

Don't install click handers on the section titles in non-tabview mode.  And
stop making them clickable.  It was pointless, since there's no tab to
switch, so it didn't do anything.  It's confusing to have clickable buttons
that don't do anything.