fix: Fix style of tab headers in BAM
Path wasn't right for the css block that was supposed to adjust their margin.

The extra gap between elements that dialogv2 has is unnecessary.

The was a property that was not used as it was overridden by a foundry property
that was more specific.  It didn't really do anything, so just get of it.