fix: Fix damage from agile actions, improve formatting
The way the agile actions was done messed up the columns by adding
additional agile buttons without accounting for them in the row x col
layout.  It added an extra column and squished them.

Fix this by adding the agile buttons as more actions.  This also fixes the
cut&paste, in triplicate, of the action button html.

It might also be possible to check for weapons with traits like Trip,
Disarm as well as free hands to only show the actions that can be done.

The way escape was special-cased didn't work with translations.  Stop
making it a special case and instead detect actions that should be agile
via athletics skill, since the rule on actions being agile mentions that.

Get rid of the translations of MAP levels, e.g. "-5", since they were the
same in every language.  And now they could be calculated, e.g. -3 for
Agile Grace, but that's not taken into account.

Get rid of the "()" around the MAP and adjust the css so the map buttons
don't take so much space and the number is centered inside the button.