Fix: Update Hero point reroll for new system
The reroll callback arguments changed in the PF2e system.  This should be
compatible with the old format too.

The code to abort the render hook early would skip the hero point and private
spell checks if no card collapse settings were on.  It didn't really help any,
since the check to leave early was the exact same checks done anyway.