feat: Use NPC skill proficiency for BAM actions
NPCs were considered untrained in all skills for action color, but as
trained in all skills when determining if they can use a trained skill
action.

Use the system's boolean skill.proficient value for NPC skills as if it was
untrained vs trained rank.  Trained actions will be skipped (subject to the
global show all actions setting) for non-proficient skills and the buttons
will be colored as-if untrained or trained.

Just pass the skill rank to the hbs template, not the whole skill, since
rank was the only thing used.  We don't want to modify the skill object
since that's owned by the actor.

Clean up the action filter somewhat too.