fix: Don't drop items multiple times on unconscious
The drop items on unconscious hook was running on itemCreate.  This means
it runs for every item (so 4 times, for unconscious, prone, blinded, and
off-guard conditions) for every user.

This means that multiple users on mutltiple conditions might decide to drop
items.  The debounce makes that less likely, but it's still possible.

There's an assumption that once all held items are dropped, nothing will
happen if dropHeldItemsOnBecomingUnconscious() is called again because
nothing will be held anymore.  But the database update is async and not
awaited.  It's possible the function will be called multiple times
before the database update propagates and the same items will be dropped
multiple times.

Switch the hook to preCreateItem, which is only called by the single user
who is making the actor unconscious.  So the whole multiple users acting
async problem goes away.  Trigger only when the unconscious condition is
created.  This avoid re-triggering on blinded, prone, etc.

Also use actor.inventory, which only has physical items, instead of
actor.items.  The non-physical items don't even have the isHeld property so
they'd never match.  And type annotate the actor argument, since typescript
doesn't know what actor is otherwise.  Now ts knows the results of
filtering actor.inventory must all be physical items.