feat: add GM macro for search and investigate rolls
feat: add GM macro for search and investigate rolls

Search and Investigate are two common exploration activities that require the GM to repeatedly roll secret checks for Perception and Recall Knowledge.

Making each roll individually for each room that the party enters takes time that quickly builds up during the course of a campaign.

This macro gives a GM the option to execute it once as the party is entering the room and consult the results with whatever documentation or guidelines they have for the area.