fix: BAM was supposed to have just one dialog open
This code was changed to use DialogV2.wait during v13 porting, but that's
not what it should have done.  It's supposed to save the dialog, and if
it's present the next time the macro runs, it closes it.  So what happened
was the macro didn't finish running until the dialog was closed, then it
saved the return value of the dialog, i.e. what button was clicked on, as
if it was the dialog itself.