fix: Fix and speed up spell damage name hiding
This was really inefficient, making multiple passes through the entire chat
history to find the original casting chat card of the spell.  Use findLast
to just search once and stop as soon as the casting is found.

It also had a bug where it would consider the spell to have been cast
private if it was _ever_ cast privately.  Even if the most recent casting
was public.

Note that since the chat card is modified on render, there is still a bug
that when scrolling back and rendering an old message the most recent
casting will be used, even if that casting was long after the old damage
roll.

It would have been better to do this in the preCreate hook, so it could
just be done once when the damage roll is made.