fix: Un/Mystifying NPCs didn't update combat tracker
It was using an option that doesn't exist in V13, but it didn't work
anyway.

What does work is to set the diff option to false, which forces an update
even though nothing in the combatant is changed, and then the update will
pick up the changed token name.  updateEmbeddedDocuments can be used to
update all the combatants in the combat at once.